Perform bulk cleansing before writing into sqlite database
In order to avoid second interaction with the database for cleansing, the procedure will be performed right after converting the xml file and before writing the table into the database. Additionally, the option for conversion to datetime datatypes is included into the bulk cleansing.
Black is applied
